+++ course_attendance_service/course_attendance_service/workflows/registration/adapters/registrants_repo_inmemory.py
