+++ course_attendance_service/course_attendance_service/workflows/registration/adapters/registrants_repo_zoom_test.py
@@ -1,6 +1,6 @@
 from unittest.mock import Mock
 
-from .repo_zoom import ZoomRegistrantsRepo
+from .registrants_repo_zoom import ZoomRegistrantsRepo
 from ..external.api_zoom import ZoomRestApi, ZoomRegistrantsData
 from ..core.workflows import Registrant
 
@@ -9,7 +9,7 @@ def test_repo_can_get_list_of_registrants_from_zoom_api():
     # GIVEN: the zoom api and a workhop
     zoom_api = Mock(ZoomRestApi)
     zoom_api.get_registrants = lambda status, **kwargs: {
-        'approved': [
+        "approved": [
             ZoomRegistrantsData(
                 id="some_random_chars",
                 first_name="Mo",
@@ -19,8 +19,8 @@ def test_repo_can_get_list_of_registrants_from_zoom_api():
                 status="approved",
             ),
         ],
-        'denied': [],
-        'pending': [],
+        "denied": [],
+        "pending": [],
     }[status]
 
     workshop_id = "abc"
@@ -47,7 +47,7 @@ def test_repo_can_get_correct_number_of_registrants_from_zoom_api():
     # GIVEN: the zoom api and a workhop
     zoom_api = Mock(ZoomRestApi)
     zoom_api.get_registrants = lambda status, **kwargs: {
-        'approved': [
+        "approved": [
             ZoomRegistrantsData(
                 id="some_random_chars",
                 first_name="Mo",
@@ -57,10 +57,10 @@ def test_repo_can_get_correct_number_of_registrants_from_zoom_api():
                 status="approved",
             ),
         ],
-        'denied': [],
-        'pending': [],
+        "denied": [],
+        "pending": [],
     }[status]
-    
+
     workshop_id = "abc"
 
     repo = ZoomRegistrantsRepo(zoom_api=zoom_api)
