+++ course_attendance_service/course_attendance_service/workflows/registration/core/workflows.py
@@ -2,14 +2,9 @@ from abc import ABC, abstractmethod
 from typing import Any, List, NamedTuple
 from .entities import Registrant
 from .contact_info_repo import ContactInfoRepo
+from .registrants_repo import RegistrantsRepo
 
 ###### Workflows
-
-
-class RegistrantsRepo(ABC):
-    @abstractmethod
-    def get_list_of_registrants(self, workshop_id: Any) -> List[Registrant]:
-        ...
 
 
 class RegistrantsReport(NamedTuple):
